[masterpiece, documents] Add the entire Snap project, add files
Adding the entire project, which includes the header, main footer, which contains a series of photos and navigation bar in the header, a slider and video in the main, and the rest of the site's description. There are also a series of photos and text in the footer.

Add photo, video, font, icon folders